Demonstrate how to launch from Program.cs
Many desktop applications launch from a Program.cs file.  This is how I build my own applications so I needed to figure out how to do this in WinUI.  After a bit of searching I found that Scott Jones had a demo application that did just that.  You have to modify the csproj with a special flag.  This version also initializes the DI container in this new program before it starts the app and it passes the service provider to the app through a constructor parameter.  I also added a boolean flag to the NativeMethods that can be used to enable or disable the Debugger.Break call when a native method call failure has been detected.  I noticed that I am detected a false error when the transparency of a window is set while having no border.  It works just fine and the global error value is not set but the return value from the call seems to indicate an error.