Refactor and Support Up/Down on SizeToContent
In WPF, when you want the window size to automajically adjust its size to fit the client content, its as simple as seting the SizeToContent property to WidthAndHeight.  I had emulated this somewhat in prior check-ins but I had only handled sizing down.  In this code sprint I handle the measuring required to be able to size up and down by creating a custom Grid Control that measures its children by passing Positive Infinity to get the True Desired Size out of the Grid.  Note that it currently does not calculate in the Window's non-client area such as the border and titlebar.  Since my main use for this feature is with a borderless/titleless Window, I have what I need out of it, but plan to enhance it for other peoples use.  So for now, when you click the button for SizeToContent, it will also make the window borderless.

I decided to go ahead and make a custom Window called a WpfWindow that implements an IWpfWindow interface that can be used in dependency injection.  Although I do not plan to fully emulate a WPF window, most of the features that I am adding right now are found in a WPF window.

Other than that, the rest was a refactor to gradually slide code into place.