DragMove feature is now a Behavior
See DragMoveBehavior, it is now attached to the WpfWindow classes content.  I removed a couple dead methods from ExtensionsUIElement.cs.  AddWindowBorder now resizes the Window to account for the additional non-client area BUT I do it with hard coded values.  I documented this in several locations and this will not be a permanent workaround as the goal is to make sure I account for any border or titlebar on the Window.

I updated the NativeMethods.cs MoveBy method so that it only needs the window handle instead of the handle and instance.  This simplifed the DragMove behavior.  Note that since the DragMove behavior needs the Window's handle, I obtain it by calling PInvoke.User32.GetActiveWindow.  I think this should be fine but I would prefer to somehow obtain the window that's directly associated with the AssociatedObject.  I documented this.