SizeToContent now includes Non-Client Space Requirements
Prior to this update, the SizeToContent feature, was accurate only when a
call to remove the window border/titlebar had been made.  Now the
feature works regardless of what the non-client area contains.  It does not
account for Windows with menus, that is outside the scope of its current
use.  It needs to be tested to assure that it accounts for DPI changes, I
will do that immediatly after this check-in and adjust the design if required.

There are at least 2 ways that I know of to calculate the non-client area. You
can use the AdjustWindowRectExForDpi function or request the rect for
both the client and the window and subtract the two to find the differnce.
I chose the later method, subtraction.

I updated the readme to cross another todo off the list.