Automatic Scale for DPI Changes
This check-in handles DPI changes by adjusting a scale transform that is attached
to a LayoutTransformControl that automatically wraps Content added to the
WPF Window (I plan on renaming this because its too confusing to people).  Note
that I also attempted to adjust the window size when the DPI changed but this
caused infinite recursion.  Picture this, your dragging a window from one screen
to another that has a different DPI.  Right at the boundary between those windows,
when the majority of the window is on the screen you are approaching, this will
cause the DPI change.  Now, I adjust the size of the Window.  This in turn places the
majority of the Window back on the original DPI which then causes another DPI
changed event to fire.  Its infinite recursion that cannot be stopped without
task manager.  It even locks up Visual Studio.  So, I stopped changing the window
size because scaling the content was good enough for my use.

Also in this update, I reduced the number of parameters needed on some of the
native methods.  NOTE :  that I beleive I have a flaw in how I calculate the size of
the non-client area of a Window such that it does not account for DPI differences.
Well, at least that is what my best guess is for why I am adding too much white
space onto a window when the DPI changes, if the Window has a border and
titlebar.  I plan to address that immediately.