WIP Changes to DPI Scaling of Window Content
By the time the public sees this check-in, I will have followed it by a more finalized
version.  This version will now only use the LayoutTransformControl if an actual DPI
change has occured since the Window was first activated. There is a
SetAutoDpiContentScaling function that is most likely to change in this check-in.

I have been testing the window moving it back and forth through DPI changes.
I see some oddity sometimes, such as the drop shadow not rendering.  It is
challenging to determine the cause.  I decided to give the DropShadowPanel
another shot and the problem does not exist with that control, however it does
not render nicely rounded corners like the custom one I made.  They are sqaured
off.