Updated Custom Drop Shadow to use Border.ActualHeight
Although I have not quite figured it out yet, sometimes as I am quickly moving the
window back and forth between two different DPI screens, the drop shadow
is lost.  I tried using the DropShadowPanel but it has ugly square corners while the
custom one I have does not.  In this check-in, I updated the Rectangle that is used
to create the rounded corner drop shadow to use the Border's ActualHeight and
ActualWidth properties.  This didn't solve the issue but is a better choice.