Add SettingsPage Click Handler for DPI Scale
Added a button click handler to call the new WpfWindow API to enable
or disable the Autoscale feature for when the DPI changes.  I attached
it to XAML that I am not checking in just yet because I need to try
something related to the DropShadows size as it is not scaling.