New RoundedDropShadowBehavior and Refactorings
There exists the Toolkit version of a DropShadowPanel that works great
until you want to use rounded corners in the shadow.  I was working on
my own version and have migrated it into a behavior.  In addition, it has
many logical improvements over the prior version I created.

I also created a Behaviors folder and pushed all three into it.

There is one requirement for this new shadow behavior; that its parent
be some type of panel where the required canvas and rectangle can
be automatically injected.

Known issue.  I had been working on automatic DPI adjustments.  They
were fairly stable except that the drop shadow disappeared when the
DPI changed.  I figured I would have fixed it in this version, but it is
still an issue.  My gut tells me that the shadow's size is not keeping
up with the framework element it is shadowing, but not sure yet.

Consider this a WIP.