Design Change to Properly Handle DPI changes
I had noticed that my drop shadow behavior was not rendering
after DPI changes fired.  As it turns out I was doing something
stupid and it was going unoticed until I turned my drop shadow
into a behavior.  When the DPI changed I was taking the windows
inner content (like the SettingsPage), pulling it out of the XAML
tree, wrapping it in a LayoutTransformControl and adding it back
to the XAML tree.  This of course caused the unloaded events to
fire and my OnDetached event of the behavior to fire too.

The new design just leaves the transform in the XAML tree with
a 1 to 1 scaling until a DPI change requires an update to the
transform.  To make this easier I just added a XAML file for the
WindowRootGrid control and changed the inheritence from a
Grid to a LayoutTransformControl (that happens to use a grid
for its control template).  Now I can move between DPI changes,
keep the shadow and everything renders as I hoped.  The one
sticking point is the fact that as the non-client area autoscales
via the framework, this ends up leaving additional (or less)
non-client white space.  As I talked about in other check-ins,
I cannot resize the window because the window resize causes
the DPI changed handler to fire again and gets stuck in a
recursive loop that can only be killed with TaskManager.  I still
need to open a discussion about this in the WinUI repo.

I also fixed a bunch of spelling errors in my comments.