Recursive DPI Change Issue Fixed (Maybe)
In the prior check-in I spoke about how when a DPI change fires
and the Window's non-client region resizes, I wanted to resize
the window to account for the new space requirements.  However
the Window resize was causing the DPI change event to fire again
but back on the prior DPI, then this would continue until you used
the task manager to kill it.  I just tried to force the problem again
and was unable to cause it.  I have a feeling that the design
changes in the prior check-in might have indirectly impacted this
problem.

Although I am certainly not break-dancing in celebration that it
is fixed, I am going to leave the resize calls in so I can continue
to test it.