Saved CPU Cycles by removing extra MeasureOverride
To provide the SizeToContent feature I needed to know how much
size was really needed to render all controls; DesiredSize is not
helpful for this.  Previously I calculated the true size by passing
infinity through to the base.MeasureOverride and would call it
again with the size passed via the parameter.

Not long after adding that feature I added DPI support via the UWP
toolkits LayoutTransformControl.  I just realized that this control
returns exactly the size it needs to render all content.  So there is
no need for the double  pass.

Also, I noticed an extra gap in the SettingsPage.xaml that I had not
seen before.  It was caused by assigning a Header value to the
NumberBox.  Something tells me this must have been broken in
WinUI Preview 1.  I simply removed the string assignment and now
everything lines up again.

On a side note, I hit the bug that is trying to hide from me where
the DPI changed event fires in infinite recursion causing visual
studio to freeze.  I found a related cause.  I can force the problem
by maximizing the window, then use my SizeToContent feature.
This will reduce the size of the Window, but the Window still
thinks it is maximized, now if you drag the window slowly
between two monitors that cause a DPI change it will lock up.
My guess is that I am not fully refreshing the window, or making
the wrong combo of Win32 calls that leave the Window in an
invalid state.  This should be my next challenge.