NativeMethod IsMaximized /  Call from Resize
I added an extension method to determine if a Window is currently
maximized and I am now calling it from the Window Resize method,
because I found that if you resize a Maximized window, the Window
will still think that it is maximized.  After googling for a bit there
did not seem to be a clear answer on how to resolve this.  There were
suggestions, but out of all of them, the one that worked was to
Restore the window before changing its size when it is in the maximized
state.  I do not feel comfortable about this solution, but it is a valid
workaround for now because I will need to change the design for other
reasons.

The built-in Window titlebar automatically scales upon a DPI change.
When I handle a DPI change for the Windows client content, when I
scale the size of the client area properly, it does not account for the
fact that the non-client area also scaled.  At first I thought, no
problem, I will just change the window size to take away from or add
additional space.  However, changing the window size during a DPI
change can cause inifinit DPI change events to fire, I have spoken
of this many times in prior check-ins and each time I think I found
a fix, I have been mistaken.  The last "fix" is related to the reason
for the changes in this check-in.  I noticed that when the window
was resized from a maximized state and still felt it was maximized,
this agitated the problem such that I felt it was the direct cause.
I am now realizing that I cannot code around this side effect so
I began searching to see if other people were having issues with
this.  Indeed, even the Windows Calculator removes the built in
Titlebar to replace with its own version.  Considering my personal
use case won't even have a titlebar, trying to make one work is
completely outside the scope of the pile of work I already have
in private repos.  So I am removing the non-client area
entirely and might provide a custom titlebar to place in the
client area to be controlled by the end user.  I think this is best
anyway, control the entire look of the window, its what I need
for Point of Sale apps and since you cannot swap out the
Titlebar in WinUI desktop as you can in UWP, it makes even
more sense.