Removed Temporary Hack Named Hack
I created a function early on in this project called HackForceRepaint that
I no longer see the need for thankfully.  I removed it and the two calls
to it.  It was there because when I removed the Window Titlebar, the
bottom of the Window would have a space the size of the titelbar that
you couldn't use my DragMove feature from.  That no longer seems to
be an issue (at this moment).