WIP : Gone with the Win32 Titlebar plus Refactor
Before this check-in I was providing a feature that would add the Win32
titlebar and border back into place with a click of the button.  Although
my own apps have no use for this, I was still banging my head trying
to make it work.  One of the issues with it is handling DPI changes. The
titlebar content scales at its own desire and unlike UWP apps WinUI
desktop apps do not provide a way to update the titlebar.  As I moved
between DPI changes it would shrink and expand and I had no way to
shrink or expand the window (as a workaround) because that ended up
causing infinite DPI changed events to fire and thus completely locking
the app up.

I learned that the calculator app didn't even use the Win32 titlebar and
that's all I needed to hear to ditch it.  This check-in completes a lot of
the work to replace the win32 titlebar with a custom one.   In that
effort I refactored and finally got rid of the WPF name in the Window
because I wasn't trying to mimic all WPF features and it might
mislead people to think that I can render WPF somehow.

I added some more view models, delegate commands and split up
the Window interface (now called IExtWindow and IPlatform) so
that the majoirty of the interface could remain in a netstandard
library.  The IPlatform contains the XamlRoot needed for the Dialog
Service.  SettingsPage is now MainPage.

TODO : Make maximize and restore buttons visibile only when they
are in the context to be used, among other things inside my head
somewhere.