Mod Directory.Build.targets to Prevent csproj Junk
Visual Studio has a bad habit of junkifying csproj files.  I was noticed
that once in a while a file would go missing.  Confused for a while I
eventually realized Visual Studio added a "None Remove" entry
without properly adding the file back in. Sometimes when I
was trying to set the file type while right clicking, it would just hide
the file.  With the changes that I made to the Directory.Build.props
file, there should be less of that.  The csproj files should not need
an entry to specify xaml build types or image content types any
longer.