Custom TitleBar now Displays an Icon
My Directory.Build.props file automatically includes an icon from a
single location so that every assembly uses the same one and any
consumer of these examples only needs to change it in one
location.  I now use that same icon to display an image in the custom
TitleBar that I created.