Throttle DragMove by using Pythagorean Theorem
I updated the DragMove operation to calculate Distance and then compare
it with a constant to determine if it the mouse has passed a threshold in
distance change before the Window is moved.  Although I wasn't seeing
any CPU spikes on my machine, it didn't make sense to move the window
on every single pixel change.