Updated Known Issues for DPI Change Problem
After I refactored the NativeMethods I noticed the SizeToContent
was not working well during a DPI change.  It seems to be
unpredictable.  I do not know that the refactor had anything to
do with it, or if it is just a coincidence.  I just tried to cause the
problem again for the last 5 minutes and have not been able
to do it.  I think some sleep is in order.