Improved DPI Change Handling and Refactor
Many of the changes are caused by a rename, so I will focus on
addressing what's important.  Prior to this check-in, when you
drag a Window between monitors that had a DPI change, it
would Resize the Window, causing a jarring effect.  This was
due to the fact that I was calling the SizeToContent feature,
something that had to be done when using the Win32 TitleBar.
Since my ExtWindow design no longer uses the Win32 TitleBar
I was able to remove the SizeToContent call.

Also, in the ExtWindow.xaml.cs file, I had a field called
_currentWindowState that was intended to be used  to detect
changes between Minimized, Maximized and etc so it could
fire an event.  I was not updating that field and the name
was misleading.  I changed the name to _cachedWindowState,
and assured it gets updated.  This will quite down the number
of events that fire because before today, it would fire even
when no change was detected.

All of those updates and a few other minor ones are making
the DPI change handling operate smoother.