Dependency Prop to Allow Drag on Maximized Window
Touch interactions with a control that has the DragMove behavior enabled
can lead to unpredictable results for the end user.  When I develop
full-screen KIOSK applications, I do not want the DragMove to fire when
the window is full-screen.  I added a dependency property to the
behavior so that developers can make their own decision.  For me, I
disable DragMove when the Window is maximize.