Simplify Logic for Swallowing Button DoubleTap
The ExtWindow has a Double Tap to maximize/restore feature enabled.  It
is attached to the Window.Content and would run even when an end-
user double tapped a button.  This is confusing when the user simply
meant to tap a button twice.  Prior to now, I had temporarily created a
behavior that could be attached to buttons to swallow that double tap
and prevent it from turning into a Maximize or Restore.  This was a pain
because you had to attach it to every button.  With this check-in, the
DoubleTap handler now automatically searches the UI tree under the
point that was double tapped to see if it contains a button that is
hit test visible.  If it does, the double tap is not turned into a
Window maximize or minimize.