DragMoveBehavior has Additional Properties
I added a couple more dependency properties to the DragMove behavior
that provide the ability to delay the drag by specified milliseconds
from the time the PointerDown event fires and the second one controls
the the length in pixels that must be passed before the DragMove
begins.

I did a bunch of manual testing on a touch screen and with the
mouse and made additional algorithm changes to improve the
handling.  The primary focus is on knowing when touches should be
converted to DragMove and when they shouldn't.  On a touch screen
it is often touched on an angle rather than straight down.  Touching
on an angle can cause a small pixel change as the finger slides.  That
slide can end up converting to a Dragmove when its not intended to.
These two new properties, plus the algo changes should provide
flexibility needed to tune that.