Too Many Improvements to Mention on One Line
My goals for today took me through a wild ride where I stumbled upon
many ways to improve the solution, but they were not my initial intent.
For that reason there is a lot in this check-in that would be better off
broken up had I been working in a team environment.

- [x] RoundedDropShadowBehavior CornerRadius dep prop had wrong owner type
- [x] ExtWindow was effectively calling InitializeComponent twice
- [x] TitleBar View is now a Data Template and not a Control
- [x] DataTemplates library that shows how to use x:Bind inside of it
- [x] A Recursive Data Template Selector that searches by Data Type
- [x] Large Refactor to ease dependency injection and reduce coupling
- [x] Refactor led to a renaming spree
- [x] API to Hide and Show the cursor, get/set WindowVm
- [x] Heavy commenting where applicable
- [x] Update Dependency.Build.targets to simplify globbing
- [x] Fixed issue where all XAML files were being copied to ouput.