MainPage and Related Models are fully MVVM
Prior to this check-in I had a mixture of binding functions directly to
buttons in XAML while using ICommand for others.  This check-in
makes the MainPageVm.cs class fully MVVM.  I added a new partial
class MainPageVm.Logic.cs that contains the logic associated with
the view.  The logic has no dependencies and all functions are
static.  Anytime I can remove logic from a dependence on external
state, I do the moonwalk because I am happy :).