Remove Transparency Disabled when fully Opaque
If there is zero transparency on the Window, the button to remove Window
transparency is now disabled.  Also, I removed setting Button
CommandParameter in a style because it was not working and the buttons
in the MainPageVm were already passing it in anyway.  I will probably
move to use DataTemplates for the ButtonVm.  That will make the UI
easier to build.