Update to .Net 5 preview 7 but Cautious
Note that the WinUI team recommends a different .Net 5 preview version
than what I am using.  I think they say to preview 5.  I have a lot of internal
code using the latest previews of .Net and for that reason I have been
making this demo work with the latest too.  So far there have not been
any apparent problems, but keep in mind that I am not using what the
WinUI team said to use!  So, if you add to my code and find something
you are trying with WinUI does not work, be cautious about reporting a bug
until you try it with the proper .Net SDK release.

I added a global.json to make it more clear what version of the SDK I have
been developing this against.