Comment out InitializeComWrappers and Explain
I had a call to WinRT.ComWrappersSupport.InitializeComWrappers, that I
adopted from an example Scott Jones had to show how to launch a
WinUI desktop app from a main method.  I think it was specific to other
requirements and not needed for what I have been doing.  I did leave it
in comments just in case it needs to be added back.