Refactor for LTS and Bug Fixes
Although my initial goal with this project was to provide samples for
others to learn from, I hadn't planned on using any of these libraries
directly into my own internal private projects.  Considering the
amount of effort that has gone into this so far, I want to maintain
this code as WinUI changes.  As a one-person company, that means
that it is best that I refactor the code to meet my own internal
process and continue to develop out these libraries for other people
to use.

That said, all of my projects begin with Oceanside so part of this
check-in was to rename the assemblies to match my companies
required process.

Although I had not planned on mixing in bug fixes with the
naming refactor, I happened upon some that I fixed.  First of all,
on rare cases the DPI change event was not firing.  My handler
was attached but I didn't get the event notification.  So I decided
to cut out the middleman (WinUI) and process the WndProc
event directly.  I will keep an eye on this to see if it fixed the
occasional issue.  Also, there was a bug in my new WndProc
code where custom handlers were not being called because I
had inverted a boolean flag.  Now fixed.  I am also hooking the
Window in the topmost part of the Window constructor to
help assure I catch all WM messages possible.

Note that even though I renamed the projects, I still need to
go through the dreaded process of renaming the directories
and hand editing the solution to match.