Part 2 Refactor for LTS and Bug Fixes
As stated in prior check-ins, I plan to support these libraries more than
I had initially intended and for that reason I needed to restructure the
projects so they meet my interal requirements.  The primary changes
are related to file relocations such as:
- Every Project now starts with my company name Oceanside
- Directory.Build.props updated to place all binary in root of repo
- NuGet.Config updated to place packages in root of repo
- Sample projects are in directories named sample
- Libraries and production code is under the src directory
- A Visual Studio solution is now at the root of the repo
- icon from root of repo is auto included only for exe's

Also, Native Code was split into two libs.  The win32 lib
contains the majority of the code and is not dependent
on WinUI.  The other native lib contains native code that
is dependent on the WinUI libs.  For example, needs
access to the Window class.  I changed the output of
one function that was returing Size to instead output
a width and height so that it could remain in the lib
that was not dependent on WinUI.  Everything builds
and executes as expected.  Files and projects are now
in their "final" resting location.

I do need to figure out how to make github pages
place the http content in a single subdirectory so
it does not pollute my repo root.