License Header Copyright Addition
Updated copyright in Licenseheaders to match the copyright that has
been at the root of the repo.  I am supporting this work through my
company Oceanside Software, so the name addition was the change.
Additionally, I am moving over to use editorconfig files and the
Visual Studio tool places an extra space after the license header that
I cannot control, but do not mind having.  I just needed to assure
everything was consistent.