New Function to help Determine Layout
I am debugging a few new features and had to make an explicit call to
a function that forced the container to remeasure itself to determine
how much space is required to layout all of its children. I don't know if
I will need it in the long run but for now I am checking it in.