Analyzer Cleanup
I have been beating my head against the wall because I realized that
a recent refactor left namespaces completely different than assembly
names and folder paths, but none of the analyzers is catching it!
To add insult to injury, I decided to give CodeRush a try after trying
to leave Resharper behind, but CodeRush only changed some of the
namespaces and left a bunch of references pointing to the old ones.

I then attempted to perform the update manually, went over it with
a fine-tooth comb, but the compiler was hitting a null reference
exception down in one of the WinUI project specific files.  As much
as I didn't want to, I reverted all of those changes and gave
CodeRush another try.  It pointed out some decent corrections that
I am checking in here, but none of those were namespace changes.
Long story short, I see the namespace mismatches, it is confusing
and I will correct it.