Moved IWindowNative into Base WinUI and Simplified
I had IWindowNative class in its own DLL that was going to house
NativeMethods that depended on WinUI.  I was able to remove the
WinUI dependency, but it was still coupled to CsWinRT.  Since WinUI
already brings in CSWinRT it really made since to just put it in the
WinUI base library and get rid of the DLL that only had one class.

Other small changes were coding style updates per tooling.