Minor Cleanup of the Sample App csproj
Microsoft.WinUI was being included twice in the sample app.  I also
found that I no longer have to include a direct reference to
Microsoft.Windows.CsWinRT to work-around the fact that I am using a
later version of .Net 5 than the WinUI team suggested is compatible.
With WinUI pre 1, if you  used .Net 5.0 preview 6 and maybe even 5,
it would throw a missing method exception because WinUI was calling
an API that has been taken out of the Net runtime.

I also removed an explicit setting of Nullable from the project file
because it is already set in the Directory.Build.props.