Overhaul DragMove State Machine for Corner Case
The logic for the Window DragMove behavior has worked well during the
time I have tested it in recent weeks, except for what I believe to have
been corner cases.  It was odd; once in a while when dragging around
off the debugger launch between screens, its almost as if the mouse
pointer was captured pressing down because the window would sort of
follow the mouse when it was not pressed.  In the prior design I wasn't
even truly capturing the mouse to help assure I would never get into
a state where I didn't release it, yet it seemed as if it was captured.
I did a big update and created a better state machine, finding a few
areas where corner cases would have caused issues if they hadn't
already.