Updated DragMove to Obtain Handle via Property
Inside the drag move, I was obtaining the window handle by calling
GetActiveWindow.  I had made notes to myself to change this
because I didn't feel there was any gaurantee that the proper
Window handle would be returned in all cases in a multiple window
environment.  It was best off to have the window set the handle
on the DragMoveBehavior it was associated with.