Suffixed two Methods with "NonCached" as Hint
The functions that calculate the keyboard repeat speed and delay are
based on WPF functions that cached the results.  Since WPF cached the
results I thought it would be best to remind the caller that they too
should not put the function in a control flow that called it often as it
would add overhead that was not necessary.